# docs/swagger/swagger.yaml
basePath: /
definitions:
  models.TokenResponse:
    properties:
      access_token:
        type: string
      code:
        type: string
      datetime:
        type: integer
      expires_in:
        type: integer
      message:
        type: string
      modulos:
        type: integer
      scope:
        type: string
      token_type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Basic authentication with client_id as username and client_secret as
    password (for token generation only).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: WSICRMREST API
  version: 1.26.4.27
paths:
  /connect/v1/token:
    get:
      consumes:
      - application/json
      description: |-
        Generates a JWT access token based on client_id and client_secret provided via Basic Auth

        **Security Features:**
        - Client secret comparison uses constant-time algorithm (prevents timing attacks)
        - Rate limiting applied (default: 60 requests/minute)
        - All attempts are logged in WSAPLLOGTOKEN table
        - Failed attempts logged with error details

        **Response Headers:**
        - X-RateLimit-Limit-Minute: Maximum requests per minute
        - X-RateLimit-Limit-Hour: Maximum requests per hour
        - X-RateLimit-Remaining-Minute: Remaining requests this minute
        - X-RateLimit-Remaining-Hour: Remaining requests this hour

        **Example:**
        ```bash
        curl -X GET "https://api.example.com/connect/v1/token" -H "Authorization: Basic $(echo -n 'client_id:client_secret' | base64)" -H "Grant_type: client_credentials"
        ```
      parameters:
      - description: Basic Auth (Base64 encoded client_id:client_secret)
        in: header
        name: Authorization
        required: true
        type: string
      - description: Must be 'client_credentials'
        in: header
        name: Grant_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token generated successfully
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "401":
          description: Invalid credentials or grant type
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "403":
          description: Database connection error
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "409":
          description: Invalid client_id or disabled application
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "429":
          description: Rate limit exceeded
          schema:
            type: object
        "500":
          description: Error generating token
          schema:
            $ref: '#/definitions/models.TokenResponse'
      security:
      - BasicAuth: []
      summary: Generate JWT token
      tags:
      - Authentication
...
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...